deps: update V8 to 11.3.244.4
PR-URL: https://github.com/nodejs/node/pull/47251

+++ deps/v8/src/execution/stack-guard.cc
@@ -81,14 +81,14 @@ void StackGuard::PushInterruptsScope(InterruptsScope* scope) {
   DCHECK_NE(scope->mode_, InterruptsScope::kNoop);
   if (scope->mode_ == InterruptsScope::kPostponeInterrupts) {
     // Intercept already requested interrupts.
-    intptr_t intercepted =
+    uint32_t intercepted =
         thread_local_.interrupt_flags_ & scope->intercept_mask_;
     scope->intercepted_flags_ = intercepted;
     thread_local_.interrupt_flags_ &= ~intercepted;
   } else {
     DCHECK_EQ(scope->mode_, InterruptsScope::kRunInterrupts);
     // Restore postponed interrupts.
-    int restored_flags = 0;
+    uint32_t restored_flags = 0;
     for (InterruptsScope* current = thread_local_.interrupt_scopes_;
          current != nullptr; current = current->prev_) {
       restored_flags |= (current->intercepted_flags_ & scope->intercept_mask_);
@@ -116,7 +116,7 @@ void StackGuard::PopInterruptsScope() {
     DCHECK_EQ(top->mode_, InterruptsScope::kRunInterrupts);
     // Postpone existing interupts if needed.
     if (top->prev_) {
-      for (int interrupt = 1; interrupt < ALL_INTERRUPTS;
+      for (uint32_t interrupt = 1; interrupt < ALL_INTERRUPTS;
            interrupt = interrupt << 1) {
         InterruptFlag flag = static_cast<InterruptFlag>(interrupt);
         if ((thread_local_.interrupt_flags_ & flag) &&
